+++ scripts/utility/summons.js
@@ -1,6 +1,6 @@
 import {chris} from '../helperFunctions.js';
 import {socket} from '../module.js';
-async function spawn(sourceActors, updates, duration, originItem, useActorOrigin = false) {
+async function spawn(sourceActors, updates, duration, originItem, useActorOrigin = false, groupInitiative = false) {
     async function effectMacro () {
         let summons = effect.flags['chris-premades']?.summons?.ids[effect.label];
         if (!summons) return;
@@ -56,14 +56,18 @@ async function spawn(sourceActors, updates, duration, originItem, useActorOrigin
     if (!updates) updates = {};
     setProperty(updates, 'embedded.ActiveEffect.Summoned Creature', effectData);
     let options = {
-        'controllingActor': originItem.actor
+        'controllingActor': originItem.actor,
+        'crosshairs': {
+            'interval': -1
+        }
     };
     let summonsIds = effect.flags['chris-premades']?.summons?.ids[originItem.name] ?? [];
     let overwriteInitiative = chris.getConfiguration(originItem, 'overwriteInitiative');
+    let groupInitiativeValue;
     for (let i of sourceActors) {
         let tokenDocument = await i.getTokenDocument();
         let updates2 = duplicate(updates);
-        if (originItem.actor.flags['chris-premades']?.feature?.undeadThralls && originItem.system.school === 'nec') {
+        if (originItem.actor.flags['chris-premades']?.feature?.undeadThralls && originItem.system.school === 'nec') { // Undead Thralls automation
             let wizardLevels = originItem.actor.classes.wizard?.system?.levels;
             if (wizardLevels) {
                 setProperty(updates2, 'actor.system.attributes.hp.formula', i.system.attributes.hp.formula + ' + ' + wizardLevels);
@@ -71,6 +75,7 @@ async function spawn(sourceActors, updates, duration, originItem, useActorOrigin
                 setProperty(updates2, 'actor.system.bonuses.rwak.damage', originItem.actor.system.attributes.prof);
             }
         }
+        options.crosshairs.interval = tokenDocument.width % 2 === 0 ? 1 : -1;
         let spawnedTokens = await warpgate.spawn(tokenDocument, updates2, {}, options);
         if (!spawnedTokens) return;
         let spawnedToken = game.canvas.scene.tokens.get(spawnedTokens[0]);
@@ -80,7 +85,15 @@ async function spawn(sourceActors, updates, duration, originItem, useActorOrigin
             let casterCombatant = game.combat.combatants.contents.find(combatant => combatant.actorId === originItem.actor.id);
             if (casterCombatant) {
                 let initiative;
-                if (game.settings.get('chris-premades', 'Tasha Initiative') != overwriteInitiative) {
+                if (groupInitiative) {
+                    if (groupInitiativeValue) {
+                        await socket.executeAsGM('createCombatant', spawnedToken.id, spawnedToken.actor.id, canvas.scene.id, groupInitiativeValue);
+                    } else {
+                        await socket.executeAsGM('createCombatant', spawnedToken.id, spawnedToken.actor.id, canvas.scene.id, null);
+                        await spawnedToken.actor.rollInitiative();
+                        groupInitiativeValue = spawnedToken.actor.initiative;
+                    }
+                } else if (game.settings.get('chris-premades', 'Tasha Initiative') != overwriteInitiative) {
                     initiative = casterCombatant.initiative - 0.01;
                     await socket.executeAsGM('createCombatant', spawnedToken.id, spawnedToken.actor.id, canvas.scene.id, initiative)
                 } else {